pyproject.toml
[project]
name = "myagents"
version = "0.1.0"
description = "Simplified Security Design Review Agent"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.64.0",
    "openai>=1.102.0",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/myagents"]
